Fixed dependencies
TernaryPoll and Poll classes moved TO the storagelibrary FROM the :app directory
