+++ app/build.gradle
@@ -29,4 +29,5 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     implementation project(path: ':smslibrary')
+    compile project(path: ':storagelibrary')
 }
