+++ storagelibrary/src/main/java/com/eis0/storagelibrary/Poll.java
@@ -1,36 +1,36 @@
-package com.eis0.library_demo;
+package com.eis0.storagelibrary;
 
 import com.eis0.smslibrary.SMSPeer;
 
 /**
  * Class to extend to create a poll
  */
-abstract class Poll {
+public abstract class Poll {
 
 
     /**
      * @return Int representing poll ID
      */
-    abstract int getPollId();
+    protected abstract int getPollId();
 
     /**
      * Check if the user is in the poll
      * @param user the user for which the check is being requested
      * @return true if the user is in the poll, false otherwise
      */
-    abstract boolean hasUser(SMSPeer user);
+    protected abstract boolean hasUser(SMSPeer user);
 
 
     /**
      * Insert an user in the poll.
      * @param user the user to insert in the poll
      */
-    abstract void addUser(SMSPeer user);
+    protected abstract void addUser(SMSPeer user);
 
     /**
      * Return the answer of a specific user
      * @param user the user whose answer is being requested
      * @return String representing the answer
      */
-    abstract String getAnswer(SMSPeer user);
+    protected abstract String getAnswer(SMSPeer user);
 }
